+++ dashboard-1.php
@@ -25,23 +25,7 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
 	
 	<title>Lukison - Foodtown</title>
 
-	<link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Arimo:400,700,400italic">
-	<link rel="stylesheet" href="assets/css/fonts/linecons/css/linecons.css">
-	<link rel="stylesheet" href="assets/css/fonts/fontawesome/css/font-awesome.min.css">
-	<link rel="stylesheet" href="assets/css/bootstrap.css">
-	<link rel="stylesheet" href="assets/css/xenon-core.css">
-	<link rel="stylesheet" href="assets/css/xenon-forms.css">
-	<link rel="stylesheet" href="assets/css/xenon-components.css">
-	<link rel="stylesheet" href="assets/css/xenon-skins.css">
-	<link rel="stylesheet" href="assets/css/custom.css">
-
-	<script src="assets/js/jquery-1.11.1.min.js"></script>
-
-	<!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
-	<!--[if lt IE 9]>
-		<script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-		<script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-	<![endif]-->
+    <?php include ('page/style-top.tpl');?>
 	
 	
 </head>
@@ -56,328 +40,8 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
         <div class="main-content">
 
             <!-- User Info, Notifications and Menu Bar -->
-			<nav class="navbar user-info-navbar" role="navigation">
 
-				<!-- Left links for user info navbar -->
-				<ul class="user-info-menu left-links list-inline list-unstyled">
-
-					<li class="hidden-sm hidden-xs">
-						<a href="#" data-toggle="sidebar">
-							<i class="fa-bars"></i>
-						</a>
-					</li>
-
-					<li class="dropdown hover-line">
-						<a href="#" data-toggle="dropdown">
-							<i class="fa-envelope-o"></i>
-							<span class="badge badge-green">15</span>
-						</a>
-
-						<ul class="dropdown-menu messages">
-							<li>
-
-								<ul class="dropdown-menu-list list-unstyled ps-scrollbar">
-
-									<li class="active"><!-- "active" class means message is unread -->
-										<a href="#">
-											<span class="line">
-												<strong>Luc Chartier</strong>
-												<span class="light small">- yesterday</span>
-											</span>
-
-											<span class="line desc small">
-												This ain’t our first item, it is the best of the rest.
-											</span>
-										</a>
-									</li>
-
-									<li class="active">
-										<a href="#">
-											<span class="line">
-												<strong>Salma Nyberg</strong>
-												<span class="light small">- 2 days ago</span>
-											</span>
-
-											<span class="line desc small">
-												Oh he decisively impression attachment friendship so if everything.
-											</span>
-										</a>
-									</li>
-
-									<li>
-										<a href="#">
-											<span class="line">
-												Hayden Cartwright
-												<span class="light small">- a week ago</span>
-											</span>
-
-											<span class="line desc small">
-												Whose her enjoy chief new young. Felicity if ye required likewise so doubtful.
-											</span>
-										</a>
-									</li>
-
-									<li>
-										<a href="#">
-											<span class="line">
-												Sandra Eberhardt
-												<span class="light small">- 16 days ago</span>
-											</span>
-
-											<span class="line desc small">
-												On so attention necessary at by provision otherwise existence direction.
-											</span>
-										</a>
-									</li>
-
-									<!-- Repeated -->
-
-									<li class="active"><!-- "active" class means message is unread -->
-										<a href="#">
-											<span class="line">
-												<strong>Luc Chartier</strong>
-												<span class="light small">- yesterday</span>
-											</span>
-
-											<span class="line desc small">
-												This ain’t our first item, it is the best of the rest.
-											</span>
-										</a>
-									</li>
-
-									<li class="active">
-										<a href="#">
-											<span class="line">
-												<strong>Salma Nyberg</strong>
-												<span class="light small">- 2 days ago</span>
-											</span>
-
-											<span class="line desc small">
-												Oh he decisively impression attachment friendship so if everything.
-											</span>
-										</a>
-									</li>
-
-									<li>
-										<a href="#">
-											<span class="line">
-												Hayden Cartwright
-												<span class="light small">- a week ago</span>
-											</span>
-
-											<span class="line desc small">
-												Whose her enjoy chief new young. Felicity if ye required likewise so doubtful.
-											</span>
-										</a>
-									</li>
-
-									<li>
-										<a href="#">
-											<span class="line">
-												Sandra Eberhardt
-												<span class="light small">- 16 days ago</span>
-											</span>
-
-											<span class="line desc small">
-												On so attention necessary at by provision otherwise existence direction.
-											</span>
-										</a>
-									</li>
-
-								</ul>
-
-							</li>
-
-							<li class="external">
-								<a href="blank-sidebar.html">
-									<span>All Messages</span>
-									<i class="fa-link-ext"></i>
-								</a>
-							</li>
-						</ul>
-					</li>
-
-					<li class="dropdown hover-line">
-						<a href="#" data-toggle="dropdown">
-							<i class="fa-bell-o"></i>
-							<span class="badge badge-purple">7</span>
-						</a>
-
-						<ul class="dropdown-menu notifications">
-							<li class="top">
-								<p class="small">
-									<a href="#" class="pull-right">Mark all Read</a>
-									You have <strong>3</strong> new notifications.
-								</p>
-							</li>
-
-							<li>
-								<ul class="dropdown-menu-list list-unstyled ps-scrollbar">
-									<li class="active notification-success">
-										<a href="#">
-											<i class="fa-user"></i>
-
-											<span class="line">
-												<strong>New user registered</strong>
-											</span>
-
-											<span class="line small time">
-												30 seconds ago
-											</span>
-										</a>
-									</li>
-
-									<li class="active notification-secondary">
-										<a href="#">
-											<i class="fa-lock"></i>
-
-											<span class="line">
-												<strong>Privacy settings have been changed</strong>
-											</span>
-
-											<span class="line small time">
-												3 hours ago
-											</span>
-										</a>
-									</li>
-
-									<li class="notification-primary">
-										<a href="#">
-											<i class="fa-thumbs-up"></i>
-
-											<span class="line">
-												<strong>Someone special liked this</strong>
-											</span>
-
-											<span class="line small time">
-												2 minutes ago
-											</span>
-										</a>
-									</li>
-
-									<li class="notification-danger">
-										<a href="#">
-											<i class="fa-calendar"></i>
-
-											<span class="line">
-												John cancelled the event
-											</span>
-
-											<span class="line small time">
-												9 hours ago
-											</span>
-										</a>
-									</li>
-
-									<li class="notification-info">
-										<a href="#">
-											<i class="fa-database"></i>
-
-											<span class="line">
-												The server is status is stable
-											</span>
-
-											<span class="line small time">
-												yesterday at 10:30am
-											</span>
-										</a>
-									</li>
-
-									<li class="notification-warning">
-										<a href="#">
-											<i class="fa-envelope-o"></i>
-
-											<span class="line">
-												New comments waiting approval
-											</span>
-
-											<span class="line small time">
-												last week
-											</span>
-										</a>
-									</li>
-								</ul>
-							</li>
-
-							<li class="external">
-								<a href="#">
-									<span>View all notifications</span>
-									<i class="fa-link-ext"></i>
-								</a>
-							</li>
-						</ul>
-					</li>
-
-				</ul>
-
-
-				<!-- Right links for user info navbar -->
-				<ul class="user-info-menu right-links list-inline list-unstyled">
-
-					<li class="search-form"><!-- You can add "always-visible" to show make the search input visible -->
-
-						<form method="get" action="extra-search.html">
-							<input type="text" name="s" class="form-control search-field" placeholder="Type to search..." />
-
-							<button type="submit" class="btn btn-link">
-								<i class="linecons-search"></i>
-							</button>
-						</form>
-
-					</li>
-
-					<!--<li class="dropdown user-profile">
-						<a href="#" data-toggle="dropdown">
-							<img src="assets/images/user-4.png" alt="user-image" class="img-circle img-inline userpic-32" width="28" />
-							<span>
-								Arlind Nushi
-								<i class="fa-angle-down"></i>
-							</span>
-						</a>
-
-						<ul class="dropdown-menu user-profile-menu list-unstyled">
-							<li>
-								<a href="#edit-profile">
-									<i class="fa-edit"></i>
-									New Post
-								</a>
-							</li>
-							<li>
-								<a href="#settings">
-									<i class="fa-wrench"></i>
-									Settings
-								</a>
-							</li>
-							<li>
-								<a href="#profile">
-									<i class="fa-user"></i>
-									Profile
-								</a>
-							</li>
-							<li>
-								<a href="#help">
-									<i class="fa-info"></i>
-									Help
-								</a>
-							</li>
-							<li class="last">
-								<a href="extra-lockscreen.html">
-									<i class="fa-lock"></i>
-									Logout
-								</a>
-							</li>
-						</ul>
-					</li>-->
-
-					<!--<li>
-						<a href="#" data-toggle="chat">
-							<i class="fa-comments-o"></i>
-						</a>
-					</li>-->
-
-				</ul>
-
-			</nav>
+            <?php include ('page/nav-top.tpl');?>
 
 			<script type="text/javascript">
 				jQuery(document).ready(function($)
@@ -444,7 +108,7 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
                             { valueField: "2",name: "Yesterday",color: "#68b828",
                                 //label:{ visible: true, customizeText: function (){return this.valueText;}},
                             },
-                            { valueField: "3",name: "2 Day ago",color: "#fcff00",
+                            { valueField: "3",name: "2 Day ago",color: "#ffe400",
                                 //label:{ visible: true, customizeText: function (){return this.valueText;}},
                             },
 							//{ valueField: "grand", name: "Transaksi", color: "#eeeeee" },
@@ -736,204 +400,7 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
 							top: 0
 						}
 					});
-					
-					var firstMonth = {
-						dataSource: getFirstMonthViews(),
-						argumentField: 'month',
-						valueField: '2014',
-						type: 'area',
-						showMinMax: true,
-						lineColor: '#68b828',
-						minColor: '#68b828',
-						maxColor: '#7c38bc',
-						showFirstLast: false,
-					},
-					secondMonth = {
-						dataSource: getSecondMonthViews(),
-						argumentField: 'month',
-						valueField: '2014',
-						type: 'splinearea',
-						lineColor: '#68b828',
-						minColor: '#68b828',
-						maxColor: '#7c38bc',
-						pointSize: 6,
-						showMinMax: true,
-						showFirstLast: false
-					},
-					thirdMonth = {
-						dataSource: getThirdMonthViews(),
-						argumentField: 'month',
-						valueField: '2014',
-						type: 'splinearea',
-						lineColor: '#68b828',
-						minColor: '#68b828',
-						maxColor: '#7c38bc',
-						pointSize: 6,
-						showMinMax: true,
-						showFirstLast: false
-					};
-					
-					function getFirstMonthViews() {
-						return [{ month: 1, 2014: 7341 },
-						{ month: 2, 2014: 7016 },
-						{ month: 3, 2014: 7202 },
-						{ month: 4, 2014: 7851 },
-						{ month: 5, 2014: 7481 },
-						{ month: 6, 2014: 6532 },
-						{ month: 7, 2014: 6498 },
-						{ month: 8, 2014: 7191 },
-						{ month: 9, 2014: 7596 },
-						{ month: 10, 2014: 8057 },
-						{ month: 11, 2014: 8373 },
-						{ month: 12, 2014: 8636 }];
-					};
-					
-					function getSecondMonthViews() {
-						return [{ month: 1, 2014: 189742 },
-						{ month: 2, 2014: 181623 },
-						{ month: 3, 2014: 205351 },
-						{ month: 4, 2014: 245625 },
-						{ month: 5, 2014: 261319 },
-						{ month: 6, 2014: 192786 },
-						{ month: 7, 2014: 194752 },
-						{ month: 8, 2014: 207017 },
-						{ month: 9, 2014: 212665 },
-						{ month: 10, 2014: 233580 },
-						{ month: 11, 2014: 231503 },
-						{ month: 12, 2014: 232824 }];
-					};
-					
-					function getThirdMonthViews() {
-						return [{ month: 1, 2014: 398},
-						{ month: 2, 2014: 422},
-						{ month: 3, 2014: 431},
-						{ month: 4, 2014: 481},
-						{ month: 5, 2014: 551},
-						{ month: 6, 2014: 449},
-						{ month: 7, 2014: 442},
-						{ month: 8, 2014: 482},
-						{ month: 9, 2014: 517},
-						{ month: 10, 2014: 566},
-						{ month: 11, 2014: 630},
-						{ month: 12, 2014: 737}];
-					};
-					
-					
-					$('.first-month').dxSparkline(firstMonth);
-					$('.second-month').dxSparkline(secondMonth);
-					$('.third-month').dxSparkline(thirdMonth);
-					
-					
-					// Realtime Network Stats
-					var i = 0,
-						rns_values = [130,150],
-						rns2_values = [39,50],
-						realtime_network_stats = [];
-					
-					for(i=0; i<=100; i++)
-					{
-						realtime_network_stats.push({ id: i, x1: between(rns_values[0], rns_values[1]), x2: between(rns2_values[0], rns2_values[1])});
-					}
-					
-					$("#realtime-network-stats").dxChart({
-						dataSource: realtime_network_stats,
-						commonSeriesSettings: {
-							type: "area",
-							argumentField: "id"
-						},
-						series: [
-							{ valueField: "x1", name: "Packets Sent", color: '#7c38bc', opacity: .4 },
-							{ valueField: "x2", name: "Packets Received", color: '#000', opacity: .5},
-						],
-						legend: {
-							verticalAlignment: "bottom",
-							horizontalAlignment: "center"
-						},
-						commonAxisSettings: {
-							label: {
-								visible: false
-							},
-							grid: {
-								visible: true,
-								color: '#f5f5f5'
-							}
-						},
-						legend: {
-							visible: false
-						},
-						argumentAxis: {
-					        valueMarginsEnabled: false
-					    },
-						valueAxis: {
-							max: 500
-						},
-						animation: {
-							enabled: false
-						}
-					}).data('iCount', i);
-					
-					$('#network-realtime-gauge').dxCircularGauge({
-						scale: {
-							startValue: 0,
-							endValue: 200,
-							majorTick: {
-								tickInterval: 50
-							}
-						},
-						rangeContainer: {
-							palette: 'pastel',
-							width: 3,
-							ranges: [
-								{ startValue: 0, endValue: 50, color: "#7c38bc" },
-								{ startValue: 50, endValue: 100, color: "#7c38bc" },
-								{ startValue: 100, endValue: 150, color: "#7c38bc" },
-								{ startValue: 150, endValue: 200, color: "#7c38bc" },
-							],
-						},
-						value: 140,
-						valueIndicator: {
-							offset: 10,
-							color: '#7c38bc',
-							type: 'triangleNeedle',
-							spindleSize: 12
-						}
-					});
-					
-					setInterval(function(){  networkRealtimeChartTick(rns_values, rns2_values); }, 1000);
-					setInterval(function(){ networkRealtimeGaugeTick(); }, 2000);
-					setInterval(function(){ networkRealtimeMBupdate(); }, 3000);
-					
-					
-					
-					// CPU Usage Gauge
-					$("#cpu-usage-gauge").dxCircularGauge({
-						scale: {
-							startValue: 0,
-							endValue: 100,
-							majorTick: {
-								tickInterval: 25
-							}
-						},
-						rangeContainer: {
-							palette: 'pastel',
-							width: 3,
-							ranges: [
-								{ startValue: 0, endValue: 25, color: "#68b828" },
-								{ startValue: 25, endValue: 50, color: "#68b828" },
-								{ startValue: 50, endValue: 75, color: "#68b828" },
-								{ startValue: 75, endValue: 100, color: "#d5080f" },
-							],
-						},
-						value: between(30, 90),
-						valueIndicator: {
-							offset: 10,
-							color: '#68b828',
-							type: 'rectangleNeedle',
-							spindleSize: 12
-						}
-					});
-					
-					
+
 					// Resize charts
 					$(window).on('xenon.resize', function()
 					{
@@ -1224,135 +691,9 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
 
 			</footer>
 		</div>
-		
 			
-		<!-- start: Chat Section
-		<div id="chat" class="fixed">
-			
-			<div class="chat-inner">
-			
-				
-				<h2 class="chat-header">
-					<a href="#" class="chat-close" data-toggle="chat">
-						<i class="fa-plus-circle rotate-45deg"></i>
-					</a>
-					
-					Chat
-					<span class="badge badge-success is-hidden">0</span>
-				</h2>
-				
-				<script type="text/javascript">
-				// Here is just a sample how to open chat conversation box
-				jQuery(document).ready(function($)
-				{
-					var $chat_conversation = $(".chat-conversation");
-					
-					$(".chat-group a").on('click', function(ev)
-					{
-						ev.preventDefault();
-						
-						$chat_conversation.toggleClass('is-open');
-						
-						$(".chat-conversation textarea").trigger('autosize.resize').focus();
-					});
-					
-					$(".conversation-close").on('click', function(ev)
-					{
-						ev.preventDefault();
-						$chat_conversation.removeClass('is-open');
-					});
-				});
-				</script>
-				
-				
-				<div class="chat-group">
-					<strong>Favorites</strong>
-					
-					<a href="#"><span class="user-status is-online"></span> <em>Catherine J. Watkins</em></a>
-					<a href="#"><span class="user-status is-online"></span> <em>Nicholas R. Walker</em></a>
-					<a href="#"><span class="user-status is-busy"></span> <em>Susan J. Best</em></a>
-					<a href="#"><span class="user-status is-idle"></span> <em>Fernando G. Olson</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Brandon S. Young</em></a>
-				</div>
-				
-				
-				<div class="chat-group">
-					<strong>Work</strong>
-					
-					<a href="#"><span class="user-status is-busy"></span> <em>Rodrigo E. Lozano</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Robert J. Garcia</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Daniel A. Pena</em></a>
-				</div>
-				
-				
-				<div class="chat-group">
-					<strong>Other</strong>
-					
-					<a href="#"><span class="user-status is-online"></span> <em>Dennis E. Johnson</em></a>
-					<a href="#"><span class="user-status is-online"></span> <em>Stuart A. Shire</em></a>
-					<a href="#"><span class="user-status is-online"></span> <em>Janet I. Matas</em></a>
-					<a href="#"><span class="user-status is-online"></span> <em>Mindy A. Smith</em></a>
-					<a href="#"><span class="user-status is-busy"></span> <em>Herman S. Foltz</em></a>
-					<a href="#"><span class="user-status is-busy"></span> <em>Gregory E. Robie</em></a>
-					<a href="#"><span class="user-status is-busy"></span> <em>Nellie T. Foreman</em></a>
-					<a href="#"><span class="user-status is-busy"></span> <em>William R. Miller</em></a>
-					<a href="#"><span class="user-status is-idle"></span> <em>Vivian J. Hall</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Melinda A. Anderson</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Gary M. Mooneyham</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Robert C. Medina</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Dylan C. Bernal</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Marc P. Sanborn</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Kenneth M. Rochester</em></a>
-					<a href="#"><span class="user-status is-offline"></span> <em>Rachael D. Carpenter</em></a>
-				</div>
-			
-			</div>
-			
-			<!-- conversation template
-			<div class="chat-conversation">
-				
-				<div class="conversation-header">
-					<a href="#" class="conversation-close">
-						&times;
-					</a>
-					
-					<span class="user-status is-online"></span>
-					<span class="display-name">Arlind Nushi</span> 
-					<small>Online</small>
-				</div>
-				
-				<ul class="conversation-body">	
-					<li>
-						<span class="user">Arlind Nushi</span>
-						<span class="time">09:00</span>
-						<p>Are you here?</p>
-					</li>
-					<li class="odd">
-						<span class="user">Brandon S. Young</span>
-						<span class="time">09:25</span>
-						<p>This message is pre-queued.</p>
-					</li>
-					<li>
-						<span class="user">Brandon S. Young</span>
-						<span class="time">09:26</span>
-						<p>Whohoo!</p>
-					</li>
-					<li class="odd">
-						<span class="user">Arlind Nushi</span>
-						<span class="time">09:27</span>
-						<p>Do you like it?</p>
-					</li>
-				</ul>
-				
-				<div class="chat-textarea">
-					<textarea class="form-control autogrow" placeholder="Type your message"></textarea>
-				</div>
-				
-			</div>
-			
-		</div>-->
-		<!-- end: Chat Section -->
-    <div class="modal fade" id="modal-1">
+
+      <div class="modal fade" id="modal-1">
         <div class="modal-dialog">
             <div class="modal-content">
 
@@ -1371,10 +712,10 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
                 </div>
             </div>
         </div>
-    </div>
+      </div>
 
 	</div>
-    <div><?php// echo $db->week_trans_day_month();?></div>
+
 	
 	<div class="page-loading-overlay">
 		<div class="loader-2"></div>
@@ -1407,41 +748,7 @@ $someJSON='[{ month: "Januari", 2012: 546, 2013: 332, 2014: 227 },{ month: "Febr
     </script>
 
 
-	<!-- Bottom Scripts -->
-	<script src="assets/js/bootstrap.min.js"></script>
-	<script src="assets/js/TweenMax.min.js"></script>
-	<script src="assets/js/resizeable.js"></script>
-	<script src="assets/js/joinable.js"></script>
-	<script src="assets/js/xenon-api.js"></script>
-	<script src="assets/js/xenon-toggles.js"></script>
-    <script src="assets/js/moment.min.js"></script>
-
-	<!-- Imported scripts on this page -->
-	<script src="assets/js/xenon-widgets.js"></script>
-	<script src="assets/js/devexpress-web-14.1/js/globalize.min.js"></script>
-	<script src="assets/js/devexpress-web-14.1/js/dx.chartjs.js"></script>
-	<script src="assets/js/toastr/toastr.min.js"></script>
-    <script src="assets/js/daterangepicker/daterangepicker.js"></script>
-    <script src="assets/js/datepicker/bootstrap-datepicker.js"></script>
-    <script src="assets/js/timepicker/bootstrap-timepicker.min.js"></script>
-    <script src="assets/js/colorpicker/bootstrap-colorpicker.min.js"></script>
-    <script src="assets/js/select2/select2.min.js"></script>
-    <script src="assets/js/jquery-ui/jquery-ui.min.js"></script>
-    <script src="assets/js/selectboxit/jquery.selectBoxIt.min.js"></script>
-    <script src="assets/js/tagsinput/bootstrap-tagsinput.min.js"></script>
-    <script src="assets/js/typeahead.bundle.js"></script>
-    <script src="assets/js/handlebars.min.js"></script>
-    <script src="assets/js/multiselect/js/jquery.multi-select.js"></script>
-
-	<!-- JavaScripts initializations and stuff -->
-	<script src="assets/js/xenon-custom.js"></script>
-
-
-    <!-- Imported styles on this page -->
-    <link rel="stylesheet" href="assets/js/daterangepicker/daterangepicker-bs3.css">
-    <link rel="stylesheet" href="assets/js/select2/select2.css">
-    <link rel="stylesheet" href="assets/js/select2/select2-bootstrap.css">
-    <link rel="stylesheet" href="assets/js/multiselect/css/multi-select.css">
+    <?php include('page/style-bottom.tpl');?>
 
 
 </body>
