+++ page/nav-top.tpl
@@ -131,7 +131,7 @@
         </li>
 
         <li class="external">
-            <a href="../blank-sidebar.html">
+            <a href="blank-sidebar.html">
                 <span>All Messages</span>
                 <i class="fa-link-ext"></i>
             </a>
@@ -258,7 +258,7 @@
 
     <li class="search-form"><!-- You can add "always-visible" to show make the search input visible -->
 
-        <form method="get" action="../extra-search.html">
+        <form method="get" action="extra-search.html">
             <input type="text" name="s" class="form-control search-field" placeholder="Type to search..." />
 
             <button type="submit" class="btn btn-link">
