+++ page/style-bottom.tpl
@@ -1,35 +1,35 @@
 <!-- Bottom Scripts -->
-<script src="../assets/js/bootstrap.min.js"></script>
-<script src="../assets/js/TweenMax.min.js"></script>
-<script src="../assets/js/resizeable.js"></script>
-<script src="../assets/js/joinable.js"></script>
-<script src="../assets/js/xenon-api.js"></script>
-<script src="../assets/js/xenon-toggles.js"></script>
-<script src="../assets/js/moment.min.js"></script>
+<script src="assets/js/bootstrap.min.js"></script>
+<script src="assets/js/TweenMax.min.js"></script>
+<script src="assets/js/resizeable.js"></script>
+<script src="assets/js/joinable.js"></script>
+<script src="assets/js/xenon-api.js"></script>
+<script src="assets/js/xenon-toggles.js"></script>
+<script src="assets/js/moment.min.js"></script>
 
 <!-- Imported scripts on this page -->
-<script src="../assets/js/xenon-widgets.js"></script>
-<script src="../assets/js/devexpress-web-14.1/js/globalize.min.js"></script>
-<script src="../assets/js/devexpress-web-14.1/js/dx.chartjs.js"></script>
-<script src="../assets/js/toastr/toastr.min.js"></script>
-<script src="../assets/js/daterangepicker/daterangepicker.js"></script>
-<script src="../assets/js/datepicker/bootstrap-datepicker.js"></script>
-<script src="../assets/js/timepicker/bootstrap-timepicker.min.js"></script>
-<script src="../assets/js/colorpicker/bootstrap-colorpicker.min.js"></script>
-<script src="../assets/js/select2/select2.min.js"></script>
-<script src="../assets/js/jquery-ui/jquery-ui.min.js"></script>
-<script src="../assets/js/selectboxit/jquery.selectBoxIt.min.js"></script>
-<script src="../assets/js/tagsinput/bootstrap-tagsinput.min.js"></script>
-<script src="../assets/js/typeahead.bundle.js"></script>
-<script src="../assets/js/handlebars.min.js"></script>
-<script src="../assets/js/multiselect/js/jquery.multi-select.js"></script>
+<script src="assets/js/xenon-widgets.js"></script>
+<script src="assets/js/devexpress-web-14.1/js/globalize.min.js"></script>
+<script src="assets/js/devexpress-web-14.1/js/dx.chartjs.js"></script>
+<script src="assets/js/toastr/toastr.min.js"></script>
+<script src="assets/js/daterangepicker/daterangepicker.js"></script>
+<script src="assets/js/datepicker/bootstrap-datepicker.js"></script>
+<script src="assets/js/timepicker/bootstrap-timepicker.min.js"></script>
+<script src="assets/js/colorpicker/bootstrap-colorpicker.min.js"></script>
+<script src="assets/js/select2/select2.min.js"></script>
+<script src="assets/js/jquery-ui/jquery-ui.min.js"></script>
+<script src="assets/js/selectboxit/jquery.selectBoxIt.min.js"></script>
+<script src="assets/js/tagsinput/bootstrap-tagsinput.min.js"></script>
+<script src="assets/js/typeahead.bundle.js"></script>
+<script src="assets/js/handlebars.min.js"></script>
+<script src="assets/js/multiselect/js/jquery.multi-select.js"></script>
 
 <!-- JavaScripts initializations and stuff -->
-<script src="../assets/js/xenon-custom.js"></script>
+<script src="assets/js/xenon-custom.js"></script>
 
 
 <!-- Imported styles on this page -->
-<link rel="stylesheet" href="../assets/js/daterangepicker/daterangepicker-bs3.css">
-<link rel="stylesheet" href="../assets/js/select2/select2.css">
-<link rel="stylesheet" href="../assets/js/select2/select2-bootstrap.css">
-<link rel="stylesheet" href="../assets/js/multiselect/css/multi-select.css">
+<link rel="stylesheet" href="assets/js/daterangepicker/daterangepicker-bs3.css">
+<link rel="stylesheet" href="assets/js/select2/select2.css">
+<link rel="stylesheet" href="assets/js/select2/select2-bootstrap.css">
+<link rel="stylesheet" href="assets/js/multiselect/css/multi-select.css">
