+++ page/style-top.tpl
@@ -1,14 +1,14 @@
 <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Arimo:400,700,400italic">
-<link rel="stylesheet" href="../assets/css/fonts/linecons/css/linecons.css">
-<link rel="stylesheet" href="../assets/css/fonts/fontawesome/css/font-awesome.min.css">
-<link rel="stylesheet" href="../assets/css/bootstrap.css">
-<link rel="stylesheet" href="../assets/css/xenon-core.css">
-<link rel="stylesheet" href="../assets/css/xenon-forms.css">
-<link rel="stylesheet" href="../assets/css/xenon-components.css">
-<link rel="stylesheet" href="../assets/css/xenon-skins.css">
-<link rel="stylesheet" href="../assets/css/custom.css">
+<link rel="stylesheet" href="assets/css/fonts/linecons/css/linecons.css">
+<link rel="stylesheet" href="assets/css/fonts/fontawesome/css/font-awesome.min.css">
+<link rel="stylesheet" href="assets/css/bootstrap.css">
+<link rel="stylesheet" href="assets/css/xenon-core.css">
+<link rel="stylesheet" href="assets/css/xenon-forms.css">
+<link rel="stylesheet" href="assets/css/xenon-components.css">
+<link rel="stylesheet" href="assets/css/xenon-skins.css">
+<link rel="stylesheet" href="assets/css/custom.css">
 
-<script src="../assets/js/jquery-1.11.1.min.js"></script>
+<script src="assets/js/jquery-1.11.1.min.js"></script>
 
 <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
 <!--[if lt IE 9]>
