+++ page/tenant_menu_top_all.tpl
@@ -1,0 +1,17 @@
+<?php
+for ($x=1;$x <=25;$x++){
+echo '<div class="row">';
+echo    '<div class="col-sm-12">';
+echo        '<div class="chart-item-bg-2">';
+echo            '<div class="chart-item-num" data-count="this" data-from="0" data-to="0" data-suffix="#" data-duration="2">0%</div>';
+echo            '<div class="chart-item-desc">';
+echo                '<p class="col-lg-7">'.$x.'Carriage quitting securing be appetite it declared. High eyes kept so busy feel call </p>';
+echo            '</div>';
+echo            '<div class="chart-item-env">';
+echo                '<div id="tenant'.$x.'" style="width: 200px;"></div>';
+echo            '</div>';
+echo        '</div>';
+echo    '</div>';
+echo '</div>';
+}
+?>
